+++ taller2/src/logica/valueObjects/VOIngresoBoleto.java
@@ -4,7 +4,7 @@ public class VOIngresoBoleto extends VOBoleto{
 	
 	private String paseo;
 	
-	public VOIngresoBoleto(int num, String nom, int ed, long cel, double desc, String pas)
+	public VOIngresoBoleto(int num, String nom, int ed, long cel, int desc, String pas)
 	{
 		super(num, nom, ed, cel, desc);
 		paseo = pas;
